Cleanup duplicate code
