--- source/Core/BSP/Miniware/BSP_PD.c
@@ -1,12 +1,0 @@
-/*
- * BSP_PD.c
- *
- *  Created on: 21 Jul 2020
- *      Author: Ralim
- */
-
-#include "BSP_PD.h"
-#include "Model_Config.h"
-#ifdef POW_PD
-
-#endif
